feat(RangeSlider): Replace deprecated implementation
