+++ packages/react-component-library/cypress/selectors/RangeSlider.ts
