+++ packages/react-component-library/cypress/specs/RangeSlider/index.spec.ts
@@ -4,33 +4,31 @@ import { ColorNeutral200, ColorAction500 } from '@defencedigital/design-tokens'
 import selectors from '../../selectors'
 import { hexToRgb } from '../../helpers'
 
-describe('RangeSliderE', () => {
+describe('RangeSlider', () => {
   describe('the user clicks on the rail (beyond selected range)', () => {
     before(() => {
-      cy.visit(
-        '/iframe.html?id=range-slider-experimental--default&viewMode=story'
-      )
+      cy.visit('/iframe.html?id=range-slider--default&viewMode=story')
 
-      cy.get(selectors.rangeSliderE.rail).click(800, 0)
+      cy.get(selectors.RangeSlider.rail).click(800, 0)
     })
 
     it('should apply focus to the handle', () => {
-      cy.get(selectors.rangeSliderE.handle.handle).should('have.focus')
+      cy.get(selectors.RangeSlider.handle.handle).should('have.focus')
     })
 
     it('should move the handle to that position', () => {
-      cy.get(selectors.rangeSliderE.handle.value)
+      cy.get(selectors.RangeSlider.handle.value)
         .contains('33')
         .should('be.visible')
     })
 
     describe('the user clicks on the rail (within selected range)', () => {
       before(() => {
-        cy.get(selectors.rangeSliderE.rail).click(200, 0)
+        cy.get(selectors.RangeSlider.rail).click(200, 0)
       })
 
       it('should move the handle to that position', () => {
-        cy.get(selectors.rangeSliderE.handle.value)
+        cy.get(selectors.RangeSlider.handle.value)
           .contains('8')
           .should('be.visible')
       })
@@ -39,30 +37,28 @@ describe('RangeSliderE', () => {
 
   describe('thresholds', () => {
     before(() => {
-      cy.visit(
-        '/iframe.html?id=range-slider-experimental--double-threshold&viewMode=story'
-      )
+      cy.visit('/iframe.html?id=range-slider--double-threshold&viewMode=story')
     })
 
     it('renders both thresholds', () => {
-      cy.get(selectors.rangeSliderE.thresholds.betweenThresholds).should(
+      cy.get(selectors.RangeSlider.thresholds.betweenThresholds).should(
         'be.visible'
       )
-      cy.get(selectors.rangeSliderE.thresholds.aboveThresholds).should(
+      cy.get(selectors.RangeSlider.thresholds.aboveThresholds).should(
         'be.visible'
       )
     })
 
     describe.skip('the user clicks on the rail (below both thresholds)', () => {
       beforeEach(() => {
-        cy.get(selectors.rangeSliderE.rail).click(0, 0)
+        cy.get(selectors.RangeSlider.rail).click(0, 0)
       })
 
       it('should no longer render either threshold', () => {
-        cy.get(selectors.rangeSliderE.thresholds.betweenThresholds).should(
+        cy.get(selectors.RangeSlider.thresholds.betweenThresholds).should(
           'not.be.visible'
         )
-        cy.get(selectors.rangeSliderE.thresholds.aboveThresholds).should(
+        cy.get(selectors.RangeSlider.thresholds.aboveThresholds).should(
           'not.be.visible'
         )
       })
@@ -71,14 +67,12 @@ describe('RangeSliderE', () => {
 
   describe('multiple handles', () => {
     before(() => {
-      cy.visit(
-        '/iframe.html?id=range-slider-experimental--multiple-handles&viewMode=story'
-      )
+      cy.visit('/iframe.html?id=range-slider--multiple-handles&viewMode=story')
     })
 
     it('correctly sets each marker active state', () => {
       // [- - * * * * * - -]
-      cy.get(selectors.rangeSliderE.markers).should(($marker) => {
+      cy.get(selectors.RangeSlider.markers).should(($marker) => {
         expect($marker.eq(0), 'marker 1').to.have.css(
           'background-color',
           hexToRgb(ColorNeutral200)
