+++ packages/react-component-library/src/components/RangeSlider/Handle.tsx
@@ -1,58 +1,64 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { GetHandleProps, SliderItem } from 'react-compound-slider'
 
-import { useThresholdColor } from './useThresholdColor'
 import { RangeSliderPositionBag, RangeSliderValueFormatter } from '.'
+import { StyledHandleWrapper } from './partials/StyledHandleWrapper'
 import { StyledHandle } from './partials/StyledHandle'
-import { StyledPercentage } from './partials/StyledPercentage'
 import { StyledValue } from './partials/StyledValue'
 
 export interface HandleProps {
-  activeHandleID: string
   domain: ReadonlyArray<number>
   handle: SliderItem
   getHandleProps: GetHandleProps
   formatValue: RangeSliderValueFormatter
-  thresholds?: number[]
 }
 
-export const Handle: React.FC<HandleProps> = ({
-  activeHandleID,
-  domain: [min, max],
-  handle: { id, value, percent },
-  getHandleProps,
-  thresholds,
-  formatValue,
-}) => {
-  const isActive: boolean = activeHandleID === id
+export const Handle = React.forwardRef<HTMLDivElement, HandleProps>(
+  (
+    {
+      domain: [min, max],
+      handle: { id, value, percent },
+      getHandleProps,
+      formatValue,
+    },
+    ref
+  ) => {
+    const [hasFocus, setHasFocus] = useState<boolean>(false)
 
-  const positionBag: RangeSliderPositionBag = {
-    value,
-    percentage: percent,
+    const positionBag: RangeSliderPositionBag = {
+      value,
+      percentage: percent,
+    }
+
+    return (
+      <StyledHandleWrapper
+        $left={`${percent}%`}
+        $hasFocus={hasFocus}
+        {...getHandleProps(id)}
+      >
+        <StyledHandle
+          ref={ref}
+          tabIndex={0}
+          role="slider"
+          aria-label="Select range"
+          aria-valuemin={min}
+          aria-valuemax={max}
+          aria-valuenow={value}
+          data-testid="rangeslider-handle"
+          onFocus={(_) => setHasFocus(true)}
+          onBlur={(_) => setHasFocus(false)}
+        />
+        <StyledValue
+          data-testid="rangeslider-value"
+          onMouseDown={(_) =>
+            (ref as React.RefObject<HTMLDivElement>).current?.focus()
+          }
+        >
+          {formatValue(positionBag)}
+        </StyledValue>
+      </StyledHandleWrapper>
+    )
   }
-
-  return (
-    <StyledHandle
-      $isActive={isActive}
-      $thresholdColor={useThresholdColor(percent, thresholds)}
-      tabIndex={0}
-      role="slider"
-      aria-label="Select range"
-      aria-valuemin={min}
-      aria-valuemax={max}
-      aria-valuenow={value}
-      $left={`${percent}%`}
-      {...getHandleProps(id)}
-      data-testid="rangeslider-handle"
-    >
-      <StyledPercentage data-testid="rangeslider-percentage">
-        {`${Math.floor(percent)}%`}
-      </StyledPercentage>
-      <StyledValue data-testid="rangeslider-value">
-        {formatValue(positionBag)}
-      </StyledValue>
-    </StyledHandle>
-  )
-}
+)
 
 Handle.displayName = 'Handle'
