+++ packages/react-component-library/src/components/RangeSlider/constants.ts
@@ -2,9 +2,10 @@ import { selectors } from '@defencedigital/design-tokens'
 
 const { color } = selectors
 
-export const RANGE_SLIDER_BG_COLOR = color('neutral', '100')
-export const RANGE_SLIDER_TRACK_COLOR = color('action', '500')
+export const RANGE_SLIDER_BG_COLOR = color('neutral', '200')
 export const RANGE_SLIDER_HANDLE_COLOR = color('action', '500')
-export const RANGE_SLIDER_TRACK_BELOW_FIRST_THRESHOLD = color('danger', '500')
+export const RANGE_SLIDER_TRACK_COLOR = color('action', '200')
+
+export const RANGE_SLIDER_TRACK_BELOW_FIRST_THRESHOLD = color('action', '500')
 export const RANGE_SLIDER_TRACK_BETWEEN_THRESHOLDS = color('warning', '500')
-export const RANGE_SLIDER_TRACK_ABOVE_THRESHOLDS = color('action', '500')
+export const RANGE_SLIDER_TRACK_ABOVE_THRESHOLDS = color('danger', '500')
