+++ packages/react-component-library/src/components/RangeSlider/index.ts
@@ -2,5 +2,6 @@ export * from './Slider'
 export * from './Handle'
 export * from './Track'
 export * from './ThresholdTrack'
+export * from './ThresholdRail'
 export * from './Tick'
 export * from './types'
