+++ packages/react-component-library/src/components/RangeSlider/partials/StyledHandleWrapper.tsx
