+++ packages/react-component-library/src/components/RangeSlider/partials/StyledRailChunk.tsx
