+++ packages/react-component-library/src/components/RangeSlider/partials/StyledRangeSlider.tsx
@@ -1,11 +1,8 @@
 import styled, { css } from 'styled-components'
 
-import { StyledPercentage } from './StyledPercentage'
-
 interface StyledRangeSliderProps {
   $isReversed?: boolean
   $isDisabled?: boolean
-  $hasPercentage?: boolean
 }
 
 export const StyledRangeSlider = styled.div<StyledRangeSliderProps>`
@@ -19,16 +16,8 @@ export const StyledRangeSlider = styled.div<StyledRangeSliderProps>`
       opacity: 0.5;
       cursor: not-allowed;
 
-      * {
+      &&& * {
         cursor: not-allowed;
       }
     `}
-
-  ${({ $hasPercentage }) =>
-    !$hasPercentage &&
-    `
-      ${StyledPercentage} {
-        display: none;
-      }
-  `}
 `
