+++ packages/react-component-library/src/components/RangeSlider/partials/StyledTicks.tsx
@@ -4,7 +4,8 @@ export const StyledTicks = styled.div`
   div:first-of-type,
   div:last-of-type {
     div {
-      height: 16px;
+      height: 20px;
+      width: 2px;
     }
   }
 `
