+++ packages/react-component-library/src/components/RangeSlider/partials/StyledTrackChunk.tsx
