--- packages/react-component-library/src/components/RangeSliderE/index.ts
@@ -1,7 +1,0 @@
-export * from './SliderE'
-export * from './HandleE'
-export * from './TrackE'
-export * from './ThresholdTrackE'
-export * from './ThresholdRailE'
-export * from './TickE'
-export * from './types'
