--- packages/react-component-library/src/components/RangeSliderE/partials/StyledRail.tsx
@@ -1,12 +1,0 @@
-import styled from 'styled-components'
-
-export const StyledRail = styled.div`
-  position: absolute;
-  top: calc(50% + 20px);
-  left: 0;
-  display: inline-block;
-  width: 100%;
-  height: 40px;
-  transform: translateY(-100%);
-  cursor: pointer;
-`
