--- packages/react-component-library/src/components/RangeSliderE/partials/StyledTicks.tsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components'
-
-export const StyledTicks = styled.div`
-  div:first-of-type,
-  div:last-of-type {
-    div {
-      height: 20px;
-      width: 2px;
-    }
-  }
-`
