--- packages/react-component-library/src/components/RangeSliderE/types.ts
@@ -1,8 +1,0 @@
-export type RangeSliderEPositionBag = {
-  value: number
-  percentage: number
-}
-
-export type RangeSliderEValueFormatter = (
-  positionBag: RangeSliderEPositionBag
-) => string
