+++ packages/react-component-library/src/index.ts
@@ -39,7 +39,6 @@ export * from './components/Popover'
 export * from './components/ProgressIndicator'
 export * from './components/Radio'
 export * from './components/RangeSlider'
-export * from './components/RangeSliderE'
 export * from './components/SectionDivider'
 export * from './components/Select'
 export * from './components/SelectE'
